+++ plugin/methods/closeById.ts
@@ -1,7 +1,7 @@
 import {guards, modalQueue, state} from "../utils/state";
 import ModalError from "../utils/ModalError";
 import {guardToPromiseFn, runGuardQueue} from "../utils/guards";
-import Modal from "../utils/Modal";
+import Modal, {ModalPublicInterface} from "../utils/Modal";
 import {GuardFunction} from "../utils/types";
 
 /**
@@ -10,15 +10,16 @@ import {GuardFunction} from "../utils/types";
  * */
 export default function closeById(id:number) {
 
-    const indexFoRemove = modalQueue.value.findIndex((item:Modal) => item.id === id);
+    const indexRemoveElement: number
+        = modalQueue.value.findIndex((item:ModalPublicInterface) => item.id === id);
 
-    if (indexFoRemove === -1) return Promise.reject(ModalError.Undefined(id));    //Modal with id not found
+    if (indexRemoveElement === -1) return Promise.reject(ModalError.Undefined(id));    //Modal with id not found
 
     const arr = guards.get(id, "close").map((guard:GuardFunction) => guardToPromiseFn(guard, id));
 
     return runGuardQueue(arr)
         .then(() => {
-            modalQueue.value.splice(indexFoRemove, 1);
+            modalQueue.value.splice(indexRemoveElement, 1);
 
             delete state.instanceStorage[id];
             guards.delete(id)