+++ plugin/utils/Modal.ts
@@ -2,23 +2,38 @@
  * last change: 25.11.2021
  * */
 
-import {computed, ComputedRef, shallowRef} from "vue";
+import {computed, ComputedRef, reactive, shallowRef} from "vue";
 import {guards, modalQueue} from "./state";
 import {GuardFunctionWithHandle} from "./types";
 import closeById from "../methods/closeById";
 import {getInstance} from "./instances";
+import ModalError from "./ModalError";
 
-export default class Modal{
+interface EventCallbacksStorage {
+    [name: string]: (data?: any) => any//Array
+}
 
-    id:number;
-    component:any;
-    params:any;
+export interface ModalPublicInterface{
+    id: number,
+    closed: ComputedRef<boolean>,
+    close : () => Promise<void>
+
+}
+
+export default class Modal implements ModalPublicInterface{
+
+    public id:number;
+    public closed: ComputedRef;
+
+
+    protected component:any;
+    protected params:any;
     /**
-     * @description VueRef var. If modal was closed value is False
+     * @description VueRef var. If modal was closed value is TRUE
      * */
-    closed:ComputedRef;
 
-    static modalId = 0;
+    protected static modalId = 0;
+    public eventCallbacks:EventCallbacksStorage = reactive({})
 
     /**
      * Создаёт объект управления модальным окном.
@@ -33,7 +48,20 @@ export default class Modal{
 
         this.component = shallowRef(component);
         this.params = params;
-        this.closed = computed(() => !modalQueue.value.includes(this));
+
+        /**
+         * БЛЯТЬ, ПОЧЕМУ ОНО ТАК?
+         * ОТВЕТ: ЭТОТ ЕБУЧИЙ ВЬЮ, ПРИ ДОБАВЛЕНИИ В modalQueue
+         * РАСКРЫВАЕТ COMPUTED(THIS.CLOSED) И КЛАДЁТ ТУДА ТУПО ЗНАЧЕНИЕ, А НЕ
+         * COMPUTED PROP {VALUE: BOOLEAN}
+         * ЧТО ЛОГИЧНО, НО ПО УЕБАНСКИ
+         *
+         * 10.02.2022 @ЖЕНЯ, КОТОРЫЙ ЕЩЁ ПЛОХО ЗНАЕТ TS.
+         *
+         * */
+        this.closed = computed(
+            () => !modalQueue.value.find(item => item.id === this.id)
+        );
 
 
         if (component.beforeModalClose) guards.add(this.id, "close", component.beforeModalClose);
@@ -44,21 +72,39 @@ export default class Modal{
     /**
      * @description Method for closing the modal window
      * */
-    close() :Promise<void> {
+    public close() :Promise<void> {
         return closeById(this.id);
     }
 
     /**
      * @description Hook for handling modal closing
      * */
-    set onclose(func: GuardFunctionWithHandle) {
+    public set onclose(func: GuardFunctionWithHandle) {
         guards.add(this.id, "close", func);
     }
     /**
      * @description Return instance of modal component
      * */
-    get instance(){
+    public get instance(){
         return getInstance(this.id);
+    }
+
+    /**
+     * @description Event handler
+     * */
+    public on(eventName: string, callback: (data?: any) => any) {
+
+        if (typeof eventName !== 'string') throw ModalError.ModalEventNameMustBeString(eventName);
+
+        eventName = 'on' + eventName.charAt(0).toUpperCase() + eventName.slice(1);
+
+        // If eventName was added firstly
+        /*
+        if (!(eventName in this.eventCallbacks))
+            this.eventCallbacks[eventName] = []
+        */
+        this.eventCallbacks[eventName] = callback.bind(this.instance);
+
     }
 
 }