+++ web/src/pages/test/ModalLogin.vue
@@ -1,0 +1,15 @@
+<template>
+    <p>LOGIN PAGE</p>
+</template>
+
+<script>
+    export default {
+        name: "ModalLogin"
+    }
+</script>
+
+<style scoped>
+    p{
+        color: white;
+    }
+</style>